+++ libs/vre/pages/search/advanced-search/src/lib/ui/property-form/property-form-list-value/property-form-list-value.component.ts
@@ -1,4 +1,4 @@
-import { AsyncPipe } from '@angular/common';
+import { AsyncPipe, CommonModule } from '@angular/common';
 import { AfterViewInit, Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
@@ -15,6 +15,7 @@ import { PropertyFormItem } from '../../../data-access/advanced-search-store/adv
   selector: 'app-property-form-list-value',
   imports: [
     AsyncPipe,
+    CommonModule,
     MatAutocompleteModule,
     MatFormFieldModule,
     MatInputModule,
@@ -37,6 +38,8 @@ import { PropertyFormItem } from '../../../data-access/advanced-search-store/adv
       <mat-autocomplete
         #auto="matAutocomplete"
         [displayWith]="displayNode"
+        (opened)="onAutocompleteOpened()"
+        (closed)="onAutocompleteClosed()"
         (optionSelected)="onSelectionChange($event.option.value)">
         @for (node of filteredList$ | async; track trackByFn($index, node)) {
           <ng-container *ngTemplateOutlet="renderNode; context: { node: node, depth: 0 }" />
@@ -56,6 +59,7 @@ import { PropertyFormItem } from '../../../data-access/advanced-search-store/adv
     </mat-form-field>
   `,
   styleUrls: ['./property-form-list-value.component.scss'],
+  standalone: true,
 })
 export class PropertyFormListValueComponent implements OnInit, AfterViewInit, OnDestroy {
   @Input() list: ListNodeV2 | undefined = undefined;
@@ -65,7 +69,7 @@ export class PropertyFormListValueComponent implements OnInit, AfterViewInit, On
 
   destroyed: Subject<void> = new Subject<void>();
 
-  valueFilterCtrl: FormControl<string | null> = new FormControl<string | null>(null);
+  valueFilterCtrl = new FormControl<ListNodeV2 | string | null>(null);
 
   constants = Constants;
 
@@ -84,18 +88,33 @@ export class PropertyFormListValueComponent implements OnInit, AfterViewInit, On
   ngAfterViewInit(): void {
     if (this.list && this.value && typeof this.value === 'string') {
       this.selectedItem = this.findItemById(this.list, this.value);
+      if (this.selectedItem) {
+        this.valueFilterCtrl.patchValue(this.selectedItem);
+      }
     }
   }
 
   trackByFn = (index: number, item: any) => `${index}-${item.label}`;
 
-  displayNode(node: any | null): string {
-    return node ? node.label : '';
-  }
+  displayNode = (v: ListNodeV2 | string | null): string => (typeof v === 'string' ? v : (v?.label ?? ''));
 
   onSelectionChange(node: ListNodeV2) {
     this.selectedItem = node;
     this.emitValueChanged.emit(node.id);
+  }
+
+  onAutocompleteOpened() {
+    // user should get all items when opening the autocomplete for selection
+    this.valueFilterCtrl.patchValue(null);
+  }
+
+  onAutocompleteClosed() {
+    // reset the input as there are no changes
+    if (this.selectedItem) {
+      this.valueFilterCtrl.patchValue(this.selectedItem);
+    } else {
+      this.valueFilterCtrl.patchValue(null);
+    }
   }
 
   ngOnDestroy(): void {