+++ libs/vre/resource-editor/resource-editor/src/lib/properties-display/delete-resource-dialog.component.ts
@@ -29,13 +29,13 @@ import { ResourceFetcherService } from '../representations/resource-fetcher.serv
     </mat-dialog-content>
 
     <mat-dialog-actions align="end">
-      <button matButton color="primary" mat-dialog-close>
+      <button mat-button color="primary" mat-dialog-close>
         {{ 'resourceEditor.propertiesDisplay.deleteResource.noKeep' | translate }}
       </button>
       <button
         data-cy="app-delete-resource-dialog-button"
-        matButton
-        matButton="elevated"
+        mat-button
+        mat-raised-button
         [color]="'warn'"
         (click)="submit()">
         {{ 'resourceEditor.propertiesDisplay.deleteResource.yesDelete' | translate }}