+++ libs/vre/resource-editor/resource-editor/src/lib/representations/replace-file-menu-item.component.ts
@@ -9,30 +9,32 @@ import {
   ReplaceFileDialogProps,
 } from './replace-file-dialog/replace-file-dialog.component';
 
+export interface ReplaceFileDialogConfig {
+  title: string;
+  representation: string;
+}
+
 @Component({
   selector: 'app-replace-file-menu-item',
   imports: [TranslatePipe, MatButton],
   template: `
-    <button matButton="filled" (click)="openReplaceFileDialog()" data-cy="replace-file-button">
+    <button mat-flat-button (click)="openReplaceFileDialog()" data-cy="replace-file-button">
       {{ 'resourceEditor.representations.replaceFile' | translate }}
     </button>
   `,
 })
 export class ReplaceFileMenuItemComponent {
-  @Input({ required: true }) title!: string;
-  @Input({ required: true }) subtitle!: string;
+  @Input({ required: true }) dialogConfig!: ReplaceFileDialogConfig;
   @Input({ required: true }) parentResource!: ReadResource;
   @Input({ required: true }) viewContainerRef!: ViewContainerRef;
-  @Input({ required: true }) representation!: string;
 
   constructor(private readonly _dialog: MatDialog) {}
 
   openReplaceFileDialog() {
     this._dialog.open<ReplaceFileDialogComponent, ReplaceFileDialogProps>(ReplaceFileDialogComponent, {
       ...DspDialogConfig.mediumDialog({
-        title: this.title,
-        subtitle: this.subtitle,
-        representation: this.representation,
+        title: this.dialogConfig.title,
+        representation: this.dialogConfig.representation,
         resource: this.parentResource,
       }),
       viewContainerRef: this.viewContainerRef,