+++ libs/vre/resource-editor/resource-editor/src/lib/representations/text/text.component.ts
@@ -1,24 +1,24 @@
 import { Component, Input } from '@angular/core';
-import { Constants, ReadArchiveFileValue, ReadResource } from '@dasch-swiss/dsp-js';
+import { Constants, ReadResource, ReadTextFileValue } from '@dasch-swiss/dsp-js';
 import { TranslatePipe } from '@ngx-translate/core';
 import { FileRepresentationComponent } from '../file-representation.component';
 
 @Component({
-  selector: 'app-archive',
+  selector: 'app-text',
   imports: [FileRepresentationComponent, TranslatePipe],
   template: `
     <app-file-representation
       [src]="src"
       [parentResource]="parentResource"
       [dialogConfig]="{
-        title: 'resourceEditor.representations.archive.title' | translate,
+        title: 'resourceEditor.representations.text.title' | translate,
         representation: representationConstant,
       }" />
   `,
 })
-export class ArchiveComponent {
-  @Input({ required: true }) src!: ReadArchiveFileValue;
+export class TextComponent {
+  @Input({ required: true }) src!: ReadTextFileValue;
   @Input({ required: true }) parentResource!: ReadResource;
 
-  readonly representationConstant = Constants.HasArchiveFileValue;
+  readonly representationConstant = Constants.HasTextFileValue;
 }