+++ libs/vre/resource-editor/resource-editor/src/lib/resource-representation.component.ts
@@ -5,9 +5,11 @@ import { DspResource } from '@dasch-swiss/vre/shared/app-common';
 import { ArchiveComponent } from './representations/archive/archive.component';
 import { AudioComponent } from './representations/audio/audio.component';
 import { DocumentComponent } from './representations/document/document.component';
+import { PdfDocumentComponent } from './representations/document/pdf-document.component';
 import { getFileValue } from './representations/get-file-value';
 import { RepresentationConstants } from './representations/representation-constants';
 import { StillImageComponent } from './representations/still-image/still-image.component';
+import { TextComponent } from './representations/text/text.component';
 import { VideoComponent } from './representations/video/video.component';
 import { ResourceRepresentationContainerComponent } from './resource-representation-container.component';
 
@@ -26,32 +28,34 @@ import { ResourceRepresentationContainerComponent } from './resource-representat
         </app-resource-representation-container>
       }
       @case (representationConstants.document) {
-        <app-resource-representation-container>
-          <app-document
-            #document
-            [class.pdf]="fileValue.filename.split('.').pop() === 'pdf'"
-            [src]="fileValue"
-            [parentResource]="resource.res" />
-        </app-resource-representation-container>
+        @if (isPdf) {
+          <app-resource-representation-container>
+            <app-pdf-document #document [src]="fileValue" [parentResource]="resource.res" />
+          </app-resource-representation-container>
+        } @else {
+          <app-resource-representation-container height="small">
+            <app-document #document [src]="fileValue" [parentResource]="resource.res" />
+          </app-resource-representation-container>
+        }
       }
       @case (representationConstants.audio) {
-        <app-resource-representation-container [small]="true">
+        <app-resource-representation-container height="small">
           <app-audio #audio [src]="fileValue" [parentResource]="resource.res" />
         </app-resource-representation-container>
       }
       @case (representationConstants.movingImage) {
-        <app-resource-representation-container>
+        <app-resource-representation-container height="auto">
           <app-video #video [src]="fileValue" [parentResource]="resource.res" />
         </app-resource-representation-container>
       }
       @case (representationConstants.archive) {
-        <app-resource-representation-container [small]="true">
+        <app-resource-representation-container height="small">
           <app-archive #archive [src]="fileValue" [parentResource]="resource.res" />
         </app-resource-representation-container>
       }
       @case (representationConstants.text) {
-        <app-resource-representation-container [small]="true">
-          <app-archive #text [src]="fileValue" [parentResource]="resource.res" />
+        <app-resource-representation-container height="small">
+          <app-text #text [src]="fileValue" [parentResource]="resource.res" />
         </app-resource-representation-container>
       }
     }
@@ -60,7 +64,9 @@ import { ResourceRepresentationContainerComponent } from './resource-representat
     ArchiveComponent,
     AudioComponent,
     DocumentComponent,
+    PdfDocumentComponent,
     StillImageComponent,
+    TextComponent,
     VideoComponent,
     ResourceRepresentationContainerComponent,
   ],
@@ -69,6 +75,15 @@ export class ResourceRepresentationComponent implements OnChanges {
   @Input({ required: true }) resource!: DspResource;
 
   fileValue!: ReadFileValue | null;
+
+  get fileExtension() {
+    return this.fileValue?.filename.split('.').pop();
+  }
+
+  get isPdf() {
+    return this.fileExtension === 'pdf';
+  }
+
   loading = false;
   protected readonly representationConstants = RepresentationConstants;